lang: core: os: Add is_virtual function
This is a basic implementation of a detection method for whether mgmt is
running in a virtualized or containerized environment. We achieve this
by doing two types of checks: on one hand, we check if the DMI/SMBIOS
info can give us information about being on a virtualized env; on the
other, we check for the presence of known files (and its contents) can
confirm if we're inside a container. Either of these situations will
cause the function to return true, with the default case being false.
All of these checks are relatively naive and can be improved by looking
at the main inspiration for this implementation, which was systemd's own
virtualization check.